server/requirements.txt
sqlalchemy >= 1.4.17
fastapi >= 0.65.2
starlette >= 0.14.2
uvicorn >= 0.14.0
pydantic >= 1.8.
nanodb >= 0.4.3
pymysql >= 1.0.2
bcrypt >= 3.2.0
jwt >= 1.2.0
pyjwt >= 2.1.0
pillow >= 8.2.0